[cmds] Update BASIC documentation, add arch-specific host*.c files
