+++ elkscmd/basic/host.h
@@ -41,17 +41,13 @@ void host_outputChar(char c);
 void host_outputLong(long num);
 void host_outputFloat(float f);
 char *host_floatToStr(float f, char *buf);
+float host_floor(float x);
 void host_newLine();
 char *host_readLine();
 char host_getKey();
 int host_ESCPressed();
 void host_outputFreeMem(unsigned int val);
 void host_sleep(long ms);
-void host_digitalWrite(int pin,int state);
-int host_digitalRead(int pin);
-int host_analogRead(int pin);
-void host_pinMode(int pin, int mode);
-float host_floor(float x);
 
 #if DISK_FUNCTIONS
 int host_directoryListing();
@@ -59,3 +55,9 @@ int host_saveProgramToFile(char *fileName, int autoexec);
 int host_loadProgramFromFile(char *fileName);
 int host_removeFile(char *fileName);
 #endif
+
+/* architecture specific */
+void host_digitalWrite(int pin,int state);
+int host_digitalRead(int pin);
+int host_analogRead(int pin);
+void host_pinMode(int pin, int mode);
